Move types required for user-provided LiveObjects typings to ably.d.ts
Add globally defined `LiveObjectsTypes` to `intentionallyNotExported` [1]
list for `typedoc`, as typedocs does not include documentation for the
globally defined interfaces and complains that "__global.LiveObjectsTypes,
defined in ./ably.d.ts, is referenced by ably.DefaultRoot but not
included in the documentation."

[1] https://typedoc.org/documents/Options.Validation.html#intentionallynotexported
